Various changes made on the cluster
The inconsistency with tabs and spaces in mpimain.c was fixed

The method of automatically choosing the number of threads to use
while calculating the factorial is now better, and makes for much
faster results.

The execution time of the program is now shown upon completion in
the MPI version.